requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
mplfinance>=0.12.10b0
nltk>=3.8.0
shap>=0.42.0
plotly>=5.17.0
streamlit>=1.28.0
pytest>=7.4.0
python-dotenv>=1.0.0
feedparser>=6.0.0
rapidfuzz>=3.3.0
google-generativeai>=0.3.0
Pillow>=10.0.0
sentence-transformers>=2.6.0
joblib>=1.3.0
pyarrow>=14.0.0
